Refs #124. Moved footnotes testing to its own module
